requirements.txt
flask
openai
python-dotenv
pypdf
python-docx